+++ charactersheet/charactersheet/viewmodels/root.js
@@ -46,6 +46,16 @@ function RootViewModel() {
 	self.settingsViewModel = ko.observable(new SettingsViewModel());
 	self.connectionManagerViewModel = ko.observable(new ConnectionManagerViewModel());
 
+    //Tooltips
+    self.profileTooltip = ko.observable('Profile');
+    self.statsTooltip = ko.observable('Stats');
+    self.skillsTooltip = ko.observable('Skills');
+    self.spellsTooltip = ko.observable('Spells');
+    self.weaponsAndArmorTooltip = ko.observable('Weapons and Armor');
+    self.backpackTooltip = ko.observable('Backpack');
+    self.notesTooltip = ko.observable('Notes');
+    self.chatTooltip = ko.observable('Chat');
+
 	//Tab Properties
 	self.profileTabStatus = ko.pureComputed(function() {
 		if (self.playerType().visibleTabs.indexOf('profile') > -1) {
@@ -266,8 +276,19 @@ function RootViewModel() {
 		    self.connected(false);
 		});
 		Notifications.global.unload.add(self.unload);
+		HotkeysService.registerHotkey('1', self.activateProfileTab);
+		HotkeysService.registerHotkey('2', self.activateStatsTab);
+		HotkeysService.registerHotkey('3', self.activateSkillsTab);
+		HotkeysService.registerHotkey('4', self.activateSpellsTab);
+		HotkeysService.registerHotkey('5', self.activateEquipmentTab);
+		HotkeysService.registerHotkey('6', self.activateInventoryTab);
+		HotkeysService.registerHotkey('7', self.activateNotesTab);
+		HotkeysService.registerHotkey('8', function() {
+    		if(self.partyTabStatus() !== 'hidden'){
+      			self.activatePartyTab();
+    		}
+		});
 	};
-
 	/**
 	 * Signal all modules to load their data.
 	 */