+++ charactersheet/charactersheet/viewmodels/stats.js
@@ -11,6 +11,9 @@ function StatsViewModel() {
 	self.deathSaveSuccessList = ko.observableArray([]);
 	self.deathSaveFailureList = ko.observableArray([]);
 
+  var msg = 'Dexterity Bonus'
+  self.initiativeTooltip = ko.observable(msg);
+
 	self.init = function() {};
 
 	self.load = function() {
@@ -130,16 +133,37 @@ function StatsViewModel() {
 		}
 	};
 
-    /**
-     * Tells the other stats model to recalculate it's passive wisdom value.
-     */
+	/**
+	 * Reset the hit dice to an unused state up to the floor of half of the
+	 * character's level.
+	 *
+	 * This will be used primarily for long rest resets.
+	 */
+	self.resetHitDice = function(){
+		var profile = Profile.findBy(CharacterManager.activeCharacter().key()[0]);
+		var level = profile.level();
+		var restoredHitDice = Math.floor(level / 2);
+
+		ko.utils.arrayForEach(this.hitDiceList(), function(hitDice) {
+			if (hitDice.hitDiceUsed() === true){
+				if (restoredHitDice !== 0){
+						hitDice.hitDiceUsed(false);
+						restoredHitDice -= 1;
+				}
+			}
+		});
+	};
+
+  /**
+   * Tells the other stats model to recalculate it's passive wisdom value.
+   */
 	self.calculatePassiveWisdom = function() {
-        self.otherStats().updateValues();
+      self.otherStats().updateValues();
 	};
 
 	self.dataHasChanged = function() {
 	    self.otherStats().save();
 	    self.health().save();
-		Notifications.stats.changed.dispatch();
+			Notifications.stats.changed.dispatch();
 	};
 };